+++ lib/rc.dart
@@ -2,3 +2,4 @@ library dev.odroe.rc;
 
 export 'src/loader.dart';
 export 'src/rc.dart';
+export 'src/rc+environment.dart';