+++ CHANGELOG.md
@@ -1,3 +1,19 @@
-# 0.3.0
+# RC v0.3.1
 
-See readme.
+To install RC v0.3.1 with command:
+
+```bash
+dart pub add orm:0.3.1
+```
+
+Or update your `pubspec.yaml` file:
+
+```yaml
+dependencies:
+  rc: ^0.3.1
+```
+
+## What's Changed
+
+* `RC` class public `sources` field, the `sources` is a unmodifiable Map.
+* Add `RC+Environment` extension, exports environment KV.